Stops misleading output print for Mac users.
Turns out for Mac users this bit of code may not be necessary. In fact, it just produces a misleading '"OpenBCI_GUI.pde: could not open " + serial_output_portName' output. 

This removes my previous changes and is the same as the current release version except that it prevents the try statement from running for Mac users so they don't see this output.

Tested and signal is still coming through on Mac with this code!