Limit the marker values to printable characters
The previous marker code allowed non printable characters to be sent as
the marker values.  This supported markers in range [1:255].  It was
tested using BLE connection.

However, it seems there are inconsistencies in the implementation of
hub.sendCommand(String) in the ElectronHub code.  When using BLE the
string is correctly sent, however, when using wifi the String is UTF-8
encoded.  This introduces additional characters into the stream that at
minimum produces unpredictable results - and often restarts the Cyton
board.

For the time being, the best option is to limit markers to the 96
printable ascii characters.  This version of the GUI does that by
mapping 1-96 to ‘ ‘ - ‘~’ .