Differentiated Save/Load by Data Mode + Cyton Error Catch
Allows 4.0 seconds to apply User's channel settings. If the time window expires, the system halts and an error message shows at the bottom of the GUI. It's still possible to access the Cyton mode screen with a dongle and no Cyton board if you delete CytonUserSettings.json from the Settings folder.