Fix for wifi UDPx3 ('burst') mode
Starting in hub version 1.3.9, requests for the 'udpBurst' protocol from the
GUI were internally remapped to 'udp' protocol with a separate 'burst' field
set to true.  Starting at hub version 2.0.0, it appears this remapping is no
longer done, and instead the GUI command is sent straight to the wifi support
library, which is not intended to recognize 'udpBurst' as a valid protocol.
This protocol name was being passed straight through to a REST URL on the wifi
shield, which did not recognize it and returned a 404 error.  The hub
responded to this 404 error by assuming the wifi shield firmware was out of
date, throwing a RESP_ERROR_WIFI_NEEDS_UPDATE error.